move cms types to data folder
