+++ src/data/event.js
@@ -1,15 +1,5 @@
 // @flow
 export type Event = {
-  sys: {
-    id: string,
-    type: string,
-    contentType: {
-      sys: {
-        id: "event"
-      }
-    },
-    revision: number
-  },
   fields: {
     name: { [string]: string },
     eventCategories: { [string]: string[] },
@@ -27,6 +17,58 @@ export type Event = {
     ticketingUrl: { [string]: string },
     venueDetails: { [string]: string[] },
     individualEventPicture: { [string]: { sys: { id: string } } }
+  },
+  sys: {
+    id: string,
+    type: string,
+    contentType: {
+      sys: {
+        id: "event"
+      }
+    },
+    revision: number
+  }
+};
+
+export type FeaturedEvents = {
+  fields: {
+    title: { [string]: string },
+    events: { [string]: Event[] }
+  },
+  sys: {
+    id: string,
+    type: string,
+    contentType: {
+      sys: {
+        id: "featuredEvents"
+      }
+    },
+    revision: number
+  }
+};
+
+export type Asset = {
+  fields: {
+    title: { [string]: string },
+    file: {
+      [string]: {
+        contentType: string,
+        fileName: string,
+        url: string,
+        details: {
+          size: number,
+          image: {
+            height: number,
+            width: number
+          }
+        }
+      }
+    }
+  },
+  sys: {
+    id: string,
+    type: "Asset",
+    revision: number
   }
 };
