+++ src/integrations/cms.js
@@ -5,108 +5,20 @@ import Config from "react-native-config";
 import { createClient } from "contentful/dist/contentful.browser.min";
 import { saveCmsData, loadCmsData } from "./storage";
 import type { SavedData } from "./storage";
+import type { Event, FeaturedEvents, Asset } from "../data/event";
 
-export type CmsEntry<Fields> = {
-  fields: Fields,
-  sys: {
-    id: string,
-    type: string,
-    contentType: {
-      sys: {
-        id: string
-      }
-    },
-    revision: number
-  }
+export type CmsEntry = Event | FeaturedEvents;
+export type CmsData = {
+  entries: CmsEntry[],
+  deletedEntries: CmsEntry[],
+  assets: Asset[],
+  deletedAssets: Asset[],
+  nextSyncToken: string
 };
-
-export type Event = {
-  fields: {
-    name: { [string]: string },
-    eventCategories: { [string]: string[] },
-    startTime: { [string]: string },
-    endTime: { [string]: string },
-    location: { [string]: { lat: number, lon: number } },
-    locationName: { [string]: string },
-    eventPriceLow: { [string]: number },
-    eventPriceHigh: { [string]: number },
-    accessibilityOptions: { [string]: string[] },
-    eventDescription: { [string]: string },
-    accessibilityDetails: { [string]: string },
-    email: { [string]: string },
-    phone: { [string]: string },
-    ticketingUrl: { [string]: string },
-    venueDetails: { [string]: string[] },
-    individualEventPicture: { [string]: { sys: { id: string } } }
-  },
-  sys: {
-    id: string,
-    type: string,
-    contentType: {
-      sys: {
-        id: "event"
-      }
-    },
-    revision: number
-  }
-};
-
-export type FeaturedEvents = {
-  fields: {
-    title: { [string]: string },
-    events: { [string]: Event[] }
-  },
-  sys: {
-    id: string,
-    type: string,
-    contentType: {
-      sys: {
-        id: "featuredEvents"
-      }
-    },
-    revision: number
-  }
-};
-
-export type CmsAsset = {
-  fields: {
-    title: { [string]: string },
-    file: {
-      [string]: {
-        contentType: string,
-        fileName: string,
-        url: string,
-        details: {
-          size: number,
-          image: {
-            height: number,
-            width: number
-          }
-        }
-      }
-    }
-  },
-  sys: {
-    id: string,
-    type: "Asset",
-    revision: number
-  }
-};
-
-export type CmsEvent = Event;
-export type CmsFeaturedEvents = FeaturedEvents;
-export type CmsEntries = CmsEvent | CmsFeaturedEvents;
 
 type SyncOpts = {
   initial: boolean,
   nextSyncToken?: string
-};
-export type CmsData = {
-  entries: CmsEntries[],
-  deletedEntries: CmsEntries[],
-  assets: CmsAsset[],
-  deletedAssets: CmsAsset[],
-  nextSyncToken: string
 };
 type Client = {
   sync: SyncOpts => Promise<CmsData>
